PREAPPS-825: Breakout types into their corrosponding directories.
Annotate code in OfflineQueueLink.
Remove redundant deduplication/cancellations.
