+++ src/apollo/offline-queue-link/errors.ts
@@ -1,0 +1,6 @@
+export class DedupedByQueueError extends Error {
+	constructor() {
+		super('Operation was deduplicated by offline-queue-link.');
+		Object.defineProperty(this, 'name', { value: 'DedupedByQueueError' });
+	}
+}
