+++ src/apollo/sync-offline-operations/types.ts
@@ -1,0 +1,7 @@
+import { StorageProvider } from '../offline-queue-link/types';
+
+export interface SyncOfflineOperationsOptions {
+	apolloClient: any;
+	storage: StorageProvider;
+	storeKey?: string;
+}
